refactor(types): extract Branding
