+++ dotcom-rendering/src/types/branding.ts
@@ -1,0 +1,14 @@
+type BrandingLogo = {
+	src: string;
+	link: string;
+	label: string;
+	dimensions: { width: number; height: number };
+};
+
+export interface Branding {
+	brandingType?: { name: string };
+	sponsorName: string;
+	logo: BrandingLogo;
+	aboutThisLink: string;
+	logoForDarkBackground?: BrandingLogo;
+}
