+++ dotcom-rendering/src/web/components/ArticleMeta.tsx
@@ -10,6 +10,7 @@ import {
 } from '@guardian/source-foundations';
 import { StraightLines } from '@guardian/source-react-components-development-kitchen';
 import { getSoleContributor } from '../../lib/byline';
+import type { Branding as BrandingType } from '../../types/branding';
 import { interactiveLegacyClasses } from '../layouts/lib/interactiveLegacyStyling';
 import { decidePalette } from '../lib/decidePalette';
 import { Avatar } from './Avatar';
@@ -30,7 +31,7 @@ type Props = {
 	tags: TagType[];
 	primaryDateline: string;
 	secondaryDateline: string;
-	branding?: Branding;
+	branding?: BrandingType;
 	discussionApiUrl: string;
 	shortUrlId: string;
 	isCommentable: boolean;
