+++ dotcom-rendering/src/web/components/Carousel.importable.tsx
@@ -11,6 +11,7 @@ import {
 } from '@guardian/source-foundations';
 import libDebounce from 'lodash.debounce';
 import { useEffect, useRef, useState } from 'react';
+import type { Branding } from '../../types/branding';
 import type { TrailType } from '../../types/trails';
 import { decidePalette } from '../lib/decidePalette';
 import { formatAttrString } from '../lib/formatAttrString';
